+++ src/main/java/com/example/demo/domain/ingredientdomain/IngredientProjection.java
@@ -1,5 +1,6 @@
 package com.example.demo.domain.ingredientdomain;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface IngredientProjection {
@@ -7,6 +8,6 @@ public interface IngredientProjection {
 
     String getName();
 
-    public Double getPrice();
+    public BigDecimal getPrice();
 
 }
